+++ src/app/(home)/components/MapSection/Markers.tsx
@@ -1,5 +1,5 @@
 import { useGetPlacesAround } from '@/apis/place';
-import { useCoordinatesStore } from '@/store/coordinateState';
+import { useCoordinatesStore } from '@/store/useCoordinatesStore';
 import { MapMarker } from 'react-kakao-maps-sdk';
 
 export default function Markers() {
